# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Telephone-Booth Operator API
  description: |
    Versioned REST + WebSocket API between the operator console
    (`Telephone-Booth-Operator`) and the phone-side Rust client.

    The phone client authenticates with a static Bearer API token issued from
    the operator UI. Operator browser sessions authenticate via Authentik OIDC
    (cookie session). Native mobile apps in
    `Telephone-Booth-Operator-Mobile` authenticate with Authentik PKCE and send
    the resulting access token as `Authorization: Bearer <jwt>`
    (`operatorBearer` security scheme).
  version: 1.0.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8787
    description: Local dev
  - url: https://operator.example.com
    description: Production
...
paths:
...
  /v1/questions:
    get:
      summary: List questions for management (drafts + active by default).
      tags: [questions]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - name: cursor
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: status
          in: query
          description: Filter to a single lifecycle state. Omit to list every non-archived question.
          schema: { $ref: "#/components/schemas/QuestionStatus" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionList" }
    post:
      summary: Create a new question from an already-uploaded audio file.
      tags: [questions]
      security:
        - operatorSession: []
        - operatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuestionCreate" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
        "404": { description: audio file not found }
        "409": { description: duplicate prompt or audio already attached }
  /v1/questions/random:
    get:
      summary: A random active question for the phone client to play.
      tags: [questions]
      security:
        - apiToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
        "404": { description: no questions available }
  /v1/questions/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Archive a question (soft-delete; stops the booth offering it).
      tags: [questions]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "204": { description: removed }
        "404": { description: not found }
  /v1/questions/{id}/activate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Activate a question so the booth offers it to callers.
      tags: [questions]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
        "404": { description: not found }
  /v1/questions/{id}/deactivate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Move a question back to draft so the booth stops offering it.
      tags: [questions]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Question" }
        "404": { description: not found }
...
components:
  securitySchemes:
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Phone-client API token issued from the operator UI.
    operatorSession:
      type: apiKey
      in: cookie
      name: __Host-booth_session
      description: HMAC-signed opaque session cookie set after OIDC login.
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: |
        Authentik-issued JWT access token. Validated against the provider's
        JWKS, issuer, audience (operator web client_id + any configured
        `OIDC_MOBILE_AUDIENCES`), and operator group / email allow-lists.
        Used by native mobile apps that authenticate via PKCE.
  schemas:
...
    Sha256:
      type: string
      pattern: "^[a-f0-9]{64}$"
    AudioRef:
      type: object
      required: [url, sha256, durationMs]
      properties:
        url: { type: string, format: uri, description: "Time-bounded read URL." }
        sha256: { $ref: "#/components/schemas/Sha256" }
        durationMs: { anyOf: [{ type: integer, minimum: 1 }, { type: "null" }] }
    Question:
      type: object
      required: [id, prompt, status, createdAt, audio]
      properties:
        id: { type: string, format: uuid }
        prompt: { type: string, minLength: 1, maxLength: 280 }
        status: { $ref: "#/components/schemas/QuestionStatus" }
        createdAt: { type: string, format: date-time }
        audio: { $ref: "#/components/schemas/AudioRef" }
    QuestionStatus:
      type: string
      enum: [draft, active, archived]
    QuestionCreate:
      type: object
      required: [prompt, audioFileId]
      properties:
        prompt: { type: string, minLength: 1, maxLength: 280 }
        audioFileId: { type: string, format: uuid }
        status:
          allOf:
            - $ref: "#/components/schemas/QuestionStatus"
          description: Defaults to draft when omitted.
    QuestionList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Question" }
        nextCursor: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
...